% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fence.collision.R
\name{fence.collision}
\alias{fence.collision}
\title{Sage grouse fence collision model}
\usage{
fence.collision(x, y, d = 3000, pad = TRUE, file.name = NULL,
  cleanup = FALSE)
}
\arguments{
\item{x}{SpatialPointsDataFrame sp class object of lek locations}

\item{y}{RasterLayer raster class object of elevation (DEM)}

\item{d}{Buffer distance (default to 3000m as defined in model)}

\item{pad}{pad d by raster resoultion ^2 to account for edge effect (TRUE/FALSE)}

\item{file.name}{Output file name, if specified results will be mosaiced into single raster}

\item{cleanup}{Delete lek-level rasters (TRUE/FALSE)}
}
\value{
raster files in tiff format. If "file.name" is specified individual lek rasters are mosaiced into a single raster.
}
\description{
Calculates sage grouse fence collision risk following methods presented in Stevens et al., (2013)
}
\details{
Uses point locations of sage-grouse lek and a 30m Digital Elevation Model with same extent and projection as lek data to build model. 
Expected estimate range 0-3 where; low risk 0-0.49, moderate risk 0.50-0.99, high risk >= 1

\itemize{ 
\item  Create a point layer representing sage-grouse lek locations.
\item  Create dissolved lek buffers 
\item  Subset elevation raster to each buffer
\item  Calculate terrain ruggedness index raster (Riley et al. 1999) 
\item  Calculate Euclidean distance to nearest lek(s)
\item  Use TRI and distance to lek and the following algebraic equation to predict collision risk (See Stevens et al., 2013 for details) 
\item  regression equation: y = 78 * exp(beta0 + beta1 * TRI + beta2 * distance), 
\item    where beta0 = -3.3254377759, beta1 = -0.2504710567, beta2 = -0.0006119843.
}
}
\examples{
\dontrun{
 library(raster)
 library(rgeos)
 library(rgdal)
 
 data(leks)  
 data(elev)

 ext <- as(extent(elev), "SpatialPolygons")
 ext <- gBuffer(ext, width = -3000)
   proj4string(ext) <- proj4string(leks)
 leks <- intersect(leks, ext)
 
 fence.collision(leks, elev, file.name = "fence_collision.tif", cleanup = TRUE)  
   fc <- raster("fence_collision.tif")
   plot(fc)
  
 # Reclassify to risk ranges and plot 
 rc <- function(x) {
   ifelse(x <= 0.49, 1, 
     ifelse(x > 0.40 & x <= 0.99, 2, 
 	  ifelse(x > 1, 3, NA)))
 }
 fc.class <- calc(fc, fun = rc, filename = "fence_collision_class.tif", overwrite = TRUE) 
   plot(fc.class, col=c("green","yellow","red"), legend = FALSE)
     points(leks, pch = 20, cex = 0.75)
     legend("topright", legend=c("Low risk","Moderate risk","High risk"),
 	       fill=c("green","yellow","red"))
 	title("Sage grouse fence collision risk - Pinedale anticline, WY")
}	
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
\references{
Riley, S.J., S.D. DeGloria and R. Elliot (1999) A terrain ruggedness index that quantifies topographic heterogeneity, Intermountain Journal of Sciences 5(1-4):23-27.

Stevens, B. S., D.E. Naugle, B. Dennis, J.W. Connelly, T. Griffiths, and K. P. Reese (2013) Mapping sage-grouse fence-collision risk: Spatially explicit models for targeting conservation implementation. Wildlife Society Bulletin, 37(2):409-415
}

