% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leks-data.R
\docType{data}
\name{leks}
\alias{leks}
\title{Sage grouse lek locations}
\format{A sp SpatialPointsDataFrame with 848 features and 2 fields:
\describe{
  \item{lekid}{unique lek ID}
  \item{counts}{number of males}
}}
\source{
\url{https://wgfd.wyo.gov/Habitat/Sage-Grouse-Management}
}
\description{
2014 Sage grouse lek locations in Wyoming - WAFWA MZII
}

